{{cookiecutter.repo_name}}/requirements.txt
asciitree==0.3.3
attrs==19.3.0
audioread==2.1.8
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
comet-git-pure==0.19.13
comet-ml==2.0.14
configobj==5.0.6; python_version >= "3.0"
cycler==0.10.0
decorator==4.4.1
docker==4.1.0
everett==1.0.2; python_version >= "3.0"
fasteners==0.15
ffmpy==0.2.2
future==0.18.2
gputil==1.4.0
gspread==3.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23; python_version < "3.8"
jams==0.3.4
joblib==0.14.0
jsonschema==3.2.0
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.30.0
matplotlib==3.1.2
mido==1.2.9
mir-eval==0.5
monotonic==1.5
more-itertools==7.2.0
musdb==0.3.1
museval==0.3.0
netifaces==0.10.9
numba==0.46.0
numcodecs==0.6.2
numpy==1.17.4
-e git+https://github.com/interactiveaudiolab/nussl.git@89790a6cd2c87c54f88a36df781d1ac866e20d4f#egg=nussl
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
pandas==0.25.3
pretty-midi==0.2.8
pyaml==19.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyparsing==2.4.5
pypiwin32==223; sys_platform == "win32" and python_version >= "3.6"
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pywin32==227; sys_platform == "win32" and python_version >= "3.6"
pyyaml==5.1.2
requests==2.22.0
resampy==0.2.2
rsa==4.0
-e git+https://github.com/pseeth/scaper.git@97cbbc479c6b72521b58923d1a73f7726d8029d9#egg=scaper
scikit-learn==0.21.3
scipy==1.3.3
simplejson==3.17.0
six==1.13.0
sklearn==0.0
sortedcontainers==2.1.0
soundfile==0.10.2
sox==1.3.7
stempeg==0.1.8
torch==1.3.1
tqdm==4.39.0
urllib3==1.25.7
websocket-client==0.56.0
wurlitzer==2.0.0
zarr==2.3.0
zipp==0.6.0; python_version < "3.8"
